* commit 'dafe99a085d0204280714f66c080fc0a7a412589': (426 commits)
  TM-17291 Fixed issue with creating override if the Favorite checkbox was clicked. Made business rule change appropriately in doing so.
  TM-17250 Dates have wrong format in the Aset Comments Detail Modal
  TM-17246 - Disabled sorting on the dependencies edit mode grid.
  TM-17275 Fix issue editing storage
  TM-17275 Remove console logs
  TM-17275 Update classes to be ignored by assets
  TM-17275 Add css classes
  TM-17275 Remove not used variable
  TM-17275 Allow set as input directive parameters the css classes to be ignored
  TM-17292 - Re-order the grid, Successors column display, On narrow content display (...) and on-hover show full content
  Fix pass to edit mode when double click Clean up code
  Merge branch 'support/4.7.x' into tales/TM-16314-add-visual-indication-that-a-view-overrides-another-view-in-asset-view-manager
  Fix pass to edit mode when double click Implement change to task details view
  Fix pass to edit mode when double click Implement change to edit view for all asset show views
  TM-17283 Changing to t.taskSpec for consistency.
  TM-17283 The HQL wouldn't accept a NULL so I used NULLIF to add the null
  Change the permission to ActionViewScript
  TM-17043 - REVERT: Update direct system asset views menu links to use override false query param
  TM-17043 - Update direct system asset views menu links to use override false query param
  TM-17260 Cast error when generation Pre-Event Checklist
  ...

+++ src/main/groovy/com/tdsops/etl/ProgressIndicator.groovy
@@ -115,7 +115,7 @@ trait ProgressIndicator {
 	 * @param currentRow
 	 * @param totalRows - number of rows in the dataSet for the current iterate loop
 	 */
-	void reportRowProgress(Integer currentRow, Integer totalRows) {
+	void reportRowProgress(Integer currentRow, Long totalRows) {
 		if (progressCallback) {
 
 			if (frequencyCounter == -1) {
